% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointillist.R
\name{jpg_df}
\alias{jpg_df}
\title{Make a data frame of jpg data}
\usage{
jpg_df(file)
}
\arguments{
\item{file}{A file path}
}
\value{
A tidy dataframe describing r, g, b and hex values
}
\description{
Make a data frame of jpg data
}

