% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointillist.R
\name{pointillise}
\alias{pointillise}
\title{Make a pointillist ggplot2 object}
\usage{
pointillise(img_df, depth = 8, density = 0.1, point_range = c(1, 2))
}
\arguments{
\item{img_df}{A tidy data frame - output from \code{img_df()}}

\item{depth}{Numeric, describing colour depth - either 8 or 24 (bit)}

\item{density}{Numeric, controls how many points are included}

\item{point_range}{Numeric vector defining the min and max point size}
}
\value{
a ggplot2 object
}
\description{
Make a pointillist ggplot2 object
}

