% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointillist.R
\name{pointillise_gif}
\alias{pointillise_gif}
\title{Animate ggplot pointillism}
\usage{
pointillise_gif(img_df, output_file, nframes, interval)
}
\arguments{
\item{img_df}{A tidy data frame - output from \code{img_df()}}

\item{output_file}{the file name of the gif}

\item{nframes}{Integer - the number of frames in the animation}

\item{interval}{Numeric - the interval between frames in seconds}
}
\value{
An animated gif file
}
\description{
Animate ggplot pointillism
}
\examples{
## ImageMagick must be installed on your system
img_path <- system.file("extdata/images", "seurat.png", package = "pointillist")

png_df <- img_df(img_path)
png_df <- colour_depth(png_df, 8)
png_df_sample <- sample_df(png_df, 0.3)
pointillise_gif(png_df_sample, "seurat.gif", 5, 0.2)
}

