% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointillist.R
\name{colour_depth}
\alias{colour_depth}
\title{Set colour depth to either 8 or 24 bit}
\usage{
colour_depth(img_df, depth)
}
\arguments{
\item{img_df}{A tidy data frame - output from \code{img_df()}}

\item{depth}{Numeric, describing colour depth - either 8 or 24 (bit)}
}
\value{
A tidy data frame with colour depth adjusted
}
\description{
Set colour depth to either 8 or 24 bit
}
\examples{
img_path <- system.file("extdata/images", "seurat.png", package = "pointillist")

png_df <- img_df(img_path)
png_df_8bit <- colour_depth(png_df, 8)
}

