% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointillist.R
\name{img_df}
\alias{img_df}
\title{Make a data frame of img data}
\usage{
img_df(file, type = NULL)
}
\arguments{
\item{file}{A file path}

\item{type}{File type - either "png" or "jpg"}
}
\value{
A tidy dataframe describing r, g, b and hex values
}
\description{
Make a data frame of img data
}

