% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointillist.R
\name{pointillise}
\alias{pointillise}
\title{Make a pointillist ggplot2 object}
\usage{
pointillise(jpg_df, depth = 8, density = 1)
}
\arguments{
\item{jpg_df}{A tidy data frame - output from \code{\link{jpg_df()}}}

\item{depth}{Numeric, describing colour depth - either 8 or 24 (bit)}

\item{density}{Numeric, controls how many points are included}
}
\value{
a ggplot2 object
}
\description{
Make a pointillist ggplot2 object
}

