% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointillist.R
\name{sample_df}
\alias{sample_df}
\title{Randomly sample a img df to reduce points}
\usage{
sample_df(img_df, density)
}
\arguments{
\item{img_df}{A tidy data frame - output from \code{img_df()}}

\item{density}{Numeric, controls how many points are included}
}
\value{
A tidy data frame with a reduced number of rows
}
\description{
Randomly sample a img df to reduce points
}

